A few changes
Rename the "i" iterator to "iSlot" for readability

Set the HUDDISABLED flag when using mana so we don't see the rage bar.

Add ManaAbility and ManaPlugin to track which slot does what ability for
ease of access.

Add the max mana pool to the HUD element.

Finish the CastAbility function.